Add static_cast

+++ src/queue/queue.cpp
@@ -124,7 +124,7 @@ void Queue::SourceLayoutChanged() {
   for (int i = 0; i < source_indexes_.count(); ++i) {
     if (!source_indexes_[i].isValid()) invalid_rows << i;
   }
-  for (int j = invalid_rows.count() - 1; j >= 0; --j) {
+  for (int j = static_cast<int>(invalid_rows.count()) - 1; j >= 0; --j) {
     const int row = invalid_rows[j];
     beginRemoveRows(QModelIndex(), row, row);
     source_indexes_.removeAt(row);
